+++ tempa/Termometer.cs
@@ -11,7 +11,7 @@ namespace CoffeeJelly.tempa
     [Serializable]
     public class TermometerAgrolog : Termometer
     {
-        public TermometerAgrolog(DateTime measurementDate, string silo, string cable, float?[] sensor) : base(measurementDate, silo, cable, sensor) { }
+        public TermometerAgrolog(DateTime measurementDate, string silo, string cable, double?[] sensor) : base(measurementDate, silo, cable, sensor) { }
 
         public override int SensorsCount => Sensors;
         public static readonly int Sensors = 7;
@@ -20,7 +20,7 @@ namespace CoffeeJelly.tempa
     [Serializable]
     public class TermometerGrainbar : Termometer
     {
-        public TermometerGrainbar(DateTime measurementDate, string silo, string cable, float?[] sensor) : base(measurementDate, silo, cable, sensor) { }
+        public TermometerGrainbar(DateTime measurementDate, string silo, string cable, double?[] sensor) : base(measurementDate, silo, cable, sensor) { }
 
         public override int SensorsCount => Sensors;
         public static readonly int Sensors = 6;
@@ -29,7 +29,7 @@ namespace CoffeeJelly.tempa
     [Serializable]
     public abstract class Termometer : ITermometer
     {
-        protected Termometer(DateTime measurementDate, string silo, string cable, float?[] sensor)
+        protected Termometer(DateTime measurementDate, string silo, string cable, double?[] sensor)
         {
             if (StringExtension.AnyNullOrEmpty(silo, cable))
                 throw new ArgumentException("All strings should have not be empty or null values.");
@@ -47,7 +47,7 @@ namespace CoffeeJelly.tempa
         /// <typeparam name="T">Тип создаваемого клаcса (должен наследовать интерфейс <see cref="ITermometer"/>).</typeparam>
         /// <returns>Возвращает экземляр созданного класса типа <typeparamref name="T"/>.</returns>
         /// <exception cref="TermometerBuildException">Исключение вызванное ошибкой создания экземпляра класса. Подробности во внутреннем исключении.</exception>
-        public static T Create<T>(DateTime measurementDate, string silo, string cable, float?[] sensor) where T : ITermometer
+        public static T Create<T>(DateTime measurementDate, string silo, string cable, double?[] sensor) where T : ITermometer
         {
             System.Threading.Monitor.Enter(_locker);
             try
@@ -67,7 +67,7 @@ namespace CoffeeJelly.tempa
         public DateTime MeasurementDate { get; }
         public string Silo { get; }
         public string Cable { get; }
-        public float?[] Sensor { get; }
+        public double?[] Sensor { get; }
         public abstract int SensorsCount { get; }
 
         private static readonly object _locker = new object();
@@ -78,7 +78,7 @@ namespace CoffeeJelly.tempa
         DateTime MeasurementDate { get;  }
         string Silo { get;  }
         string Cable { get; }
-        float?[] Sensor { get;  }
+        double?[] Sensor { get;  }
         int SensorsCount { get; }
     }
 }
